Better default tuning.
Turns out, bad tuning can make queries got a bit slower *and* use
an order or two magnitude more memory.  It doesn't take much.

I've left buffer tunability for channels to the query processor and
the doc processor.  Most of my fiddling with these things made it a
bit slower.  Increasing the number of query handlers seemed to only
slow stuff down on this box, but it's 4 HT cores, so it could be
different on a box with 8 real cores.