Use closeAll to close the channel slice.
I stripped down the original implementation which included things that
aren't strictly necessary.  The new one is slightly faster and easier
to read, but I'm including the original in the commit message beacuse
it's infromative.

// Thanks to remy_o in #go-nuts for this.
func closeAll(chans interface{}) {
	v := reflect.ValueOf(chans)
	if v.Kind() != reflect.Slice || v.Type().Elem().Kind() != reflect.Chan {
		panic("bad type")
	}
	for i, imax := 0, v.Len(); i < imax; i++ {
		v.Index(i).Close()
	}
}

Benchmarks of doing a regular range across the slice:

    BenchmarkCloseRange1    50000000            63 ns/op
    BenchmarkCloseRange3    20000000           119 ns/op
    BenchmarkCloseRange1000   100000         28899 ns/op

And then the same using the new closeAll

    BenchmarkCloseAll1      10000000           227 ns/op
    BenchmarkCloseAll3       5000000           414 ns/op
    BenchmarkCloseAll1000      20000         93062 ns/op