Also bench keying a smaller number of docs.
The test I performed here was for a rather large number of docs.  It's
good to know how it does on a variety of sizes, but I was trying to
mentally compare the 1,440 results with something I saw elsewhere on
the screen and figured it'd make more sense to just test both.