A query with no pointers would crash the server.
This closes #19 with a reason I'll repeat below:

It crashes because when there are no ptrs specified, there are no
result channels created upon which we can block. This means that we
open the db, begin walking it and immediately close it because we've
collected all the info we needed to collect.

I could make another condition to block the execution here, but
there's really no point because you just shouldn't do that.  This
change detects that at the beginning of the query and just returns a
quick error.  It panics if the part of the code that crashed ever gets
into that condition again.

Also, return an HTTP 400 when a user even tries to do this so as not
to be confusing with empty results.